+++ system/process/process_test.go
@@ -203,6 +203,49 @@ func (s *ProcessSuite) TestCalculateCPUUsage(c *C) {
 	c.Assert(CalculateCPUUsage(s1, s2, time.Second), Equals, 53.0)
 }
 
+func (s *ProcessSuite) TestCPUPriority(c *C) {
+	pid := os.Getpid()
+
+	pri, ni, err := GetCPUPriority(pid)
+	c.Assert(err, IsNil)
+	c.Assert(pri, Equals, 20)
+	c.Assert(ni, Equals, 0)
+
+	err = SetCPUPriority(pid, 2)
+	c.Assert(err, IsNil)
+
+	pri, ni, err = GetCPUPriority(pid)
+	c.Assert(err, IsNil)
+	c.Assert(pri, Equals, 22)
+	c.Assert(ni, Equals, 2)
+
+	_, _, err = GetCPUPriority(12000000)
+	c.Assert(err, NotNil)
+}
+
+func (s *ProcessSuite) TestIOPriority(c *C) {
+	pid := os.Getpid()
+
+	class, classdata, err := GetIOPriority(pid)
+	c.Assert(err, IsNil)
+	c.Assert(class, Equals, PRIO_CLASS_NONE)
+	c.Assert(classdata, Equals, 4)
+
+	err = SetIOPriority(pid, PRIO_CLASS_BEST_EFFORT, 5)
+	c.Assert(err, IsNil)
+
+	class, classdata, err = GetIOPriority(pid)
+	c.Assert(err, IsNil)
+	c.Assert(class, Equals, PRIO_CLASS_BEST_EFFORT)
+	c.Assert(classdata, Equals, 5)
+
+	_, _, err = GetIOPriority(999999)
+	c.Assert(err, NotNil)
+
+	err = SetIOPriority(999999, PRIO_CLASS_BEST_EFFORT, 5)
+	c.Assert(err, NotNil)
+}
+
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 func (s *ProcessSuite) CreateFakeProcFS(c *C, pid, file, data string) string {